requirements.txt
pydantic
flask
flask_cors
torch
torchvision
opencv-python
pyaml
numpy
Pillow
pyserial
asyncio-atexit
streamlit
streamlit-joystick
pyzmq
click